requirements.txt
# Core data science packages
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0

# Financial analysis
yfinance>=0.2.0
pandas-datareader>=0.10.0
ta-lib>=0.4.0

# Machine learning
scikit-learn>=1.3.0
statsmodels>=0.14.0
xgboost>=2.0.0
tensorflow>=2.15.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0

# Frontend: React (see frontend/package.json for JS deps)

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
duckdb>=1.0.0
pyarrow>=14.0.2
fredapi>=0.5.0
lxml>=4.9.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
